factored out region, environment, and validation methods
